Feature/jwt token manager

+++ CafeVesuviusApi/Services/IReservationRepository.cs
@@ -1,19 +1,19 @@
 using CafeVesuviusApi.DTOs;
-using CafeVesuviusApi.Models;
+using CafeVesuviusApi.Entities;
 
 namespace CafeVesuviusApi.Services
 {
     public interface IReservationRepository
     {
         Task<IEnumerable<ReservationDTO>> GetReservations();
-        Task<Reservation> GetReservation(long id);
+        Task<Reservation> GetReservation(int id);
         Task<Reservation> PostReservation(Reservation reservation);
-        Task<bool> PutReservation(long id, Reservation reservation);
-        Task<bool> DeleteReservation(long id);
+        Task<bool> PutReservation(int id, Reservation reservation);
+        Task<bool> DeleteReservation(int id);
         Task<IEnumerable<DiningTable>> GetDiningTables();
         Task<IEnumerable<DiningTable>> GetAvailableDiningTables(DateTime reservationTime);
         Task<DiningTable> PostDiningTable(DiningTable diningTable);
-        Task<bool> PutDiningTable(long id, DiningTable diningTable);
-        Task<bool> DeleteDiningTable(long id);
+        Task<bool> PutDiningTable(int id, DiningTable diningTable);
+        Task<bool> DeleteDiningTable(int id);
     }
 }
